feat: Add flags for restricted environments
Runtime-init can take optional parameters during installation to avoid
calling out to the public internet. The runtime_init_installer variable
exposes some fields to toggle the flags and change behaviour.

In addition, the setting of recommended sys db values happens in the
boot script rather than as a runtime-init pre-onboard step which should
lead to a smoother experience when onboarding.